Allow calling using `python -m ccb`
